@@ -37,97 +37,106 @@
     "help": "nx help"
   },
   "dependencies": {
-    "@angular/animations": "^8.1.1",
-    "@angular/common": "^8.1.1",
-    "@angular/compiler": "^8.1.1",
-    "@angular/core": "^8.1.1",
-    "@angular/forms": "^8.1.1",
-    "@angular/platform-browser": "^8.1.1",
-    "@angular/platform-browser-dynamic": "^8.1.1",
-    "@angular/router": "^8.1.1",
-    "@godaddy/terminus": "^4.1.2",
-    "@hapi/joi": "^15.1.0",
-    "@nestjs/common": "^6.5.2",
-    "@nestjs/core": "^6.5.2",
-    "@nestjs/graphql": "^6.4.2",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@godaddy/terminus": "^4.2.1",
+    "@golevelup/nestjs-modules": "^0.3.0",
+    "@golevelup/nestjs-testing": "^0.1.2",
+    "@nestjs/common": "^6.9.0",
+    "@nestjs/core": "^6.9.0",
+    "@nestjs/graphql": "^6.5.3",
     "@nestjs/jwt": "^6.1.1",
     "@nestjs/passport": "^6.1.0",
-    "@nestjs/platform-express": "^6.5.2",
-    "@nestjs/terminus": "^6.4.0",
-    "@nrwl/angular": "^8.2.0",
-    "apollo-angular": "^1.6.0",
-    "apollo-angular-link-http": "^1.6.0",
-    "apollo-cache-inmemory": "^1.3.2",
-    "apollo-client": "^2.6.3",
-    "apollo-link": "^1.2.11",
-    "apollo-server-express": "^2.6.9",
+    "@nestjs/platform-express": "^6.9.0",
+    "@nestjs/terminus": "^6.5.2",
+    "@nrwl/angular": "^8.7.1",
+    "apollo-angular": "^1.8.0",
+    "apollo-angular-link-http": "^1.9.0",
+    "apollo-cache-inmemory": "^1.6.3",
+    "apollo-client": "^2.6.4",
+    "apollo-link": "^1.2.13",
+    "apollo-server-express": "^2.9.9",
     "bcrypt": "^3.0.6",
     "class-transformer": "^0.2.3",
-    "class-validator": "^0.9.1",
+    "class-validator": "^0.11.0",
     "compression": "^1.7.4",
-    "connect-redis": "^3.4.1",
-    "core-js": "^3.1.4",
-    "dotenv": "^8.0.0",
+    "connect-redis": "^4.0.3",
+    "core-js": "^3.4.1",
+    "dotenv": "^8.2.0",
     "express-rate-limit": "^5.0.0",
-    "express-session": "^1.16.2",
-    "graphql": "^14.4.2",
+    "express-session": "^1.17.0",
+    "graphql": "^14.5.8",
     "graphql-tag": "^2.10.1",
-    "graphql-tools": "^4.0.5",
-    "helmet": "^3.18.0",
+    "graphql-tools": "^4.0.6",
+    "helmet": "^3.21.2",
     "mc-scribe": "^3.2.1",
     "morgan": "^1.9.1",
     "passport": "^0.4.0",
+    "passport-google-oauth": "^2.0.0",
     "passport-jwt": "^4.0.0",
-    "pg": "^7.11.0",
+    "pg": "^7.12.1",
+    "redis": "^2.8.0",
     "reflect-metadata": "^0.1.12",
-    "rxjs": "^6.5.2",
-    "zone.js": "~0.9.1"
+    "runtypes": "^4.0.3",
+    "rxjs": "^6.5.3",
+    "sqlite3": "^4.1.0",
+    "type-graphql": "^0.17.5",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.801.1",
-    "@angular/cli": "^8.1.1",
-    "@angular/compiler-cli": "^8.1.1",
-    "@angular/language-service": "^8.1.1",
-    "@commitlint/cli": "^8.1.0",
-    "@commitlint/config-conventional": "^8.0.0",
-    "@nestjs/schematics": "^6.3.0",
-    "@nestjs/testing": "^6.5.2",
-    "@nrwl/cypress": "^8.2.0",
-    "@nrwl/jest": "^8.2.0",
-    "@nrwl/nest": "^8.4.3",
-    "@nrwl/node": "^8.2.0",
-    "@nrwl/workspace": "^8.2.0",
+    "@abraham/reflection": "^0.7.0",
+    "@angular-devkit/build-angular": "^0.803.19",
+    "@angular/cli": "^8.3.19",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@commitlint/cli": "^8.2.0",
+    "@commitlint/config-conventional": "^8.2.0",
+    "@nestjs/schematics": "^6.7.2",
+    "@nestjs/testing": "^6.9.0",
+    "@nrwl/cypress": "^8.7.1",
+    "@nrwl/jest": "^8.7.1",
+    "@nrwl/nest": "^8.7.1",
+    "@nrwl/node": "^8.7.1",
+    "@nrwl/workspace": "^8.7.1",
     "@types/bcrypt": "^3.0.0",
-    "@types/compression": "0.0.36",
-    "@types/connect-redis": "0.0.10",
-    "@types/dotenv": "^6.1.1",
-    "@types/express-rate-limit": "^3.3.0",
-    "@types/express-session": "^1.15.13",
-    "@types/hapi__joi": "^15.0.3",
-    "@types/helmet": "0.0.43",
-    "@types/jest": "^24.0.15",
-    "@types/joi": "^14.3.3",
-    "@types/jquery": "^3.3.30",
-    "@types/morgan": "^1.7.35",
-    "@types/node": "^12.6.2",
-    "@types/passport": "^1.0.0",
-    "@types/passport-jwt": "^3.0.1",
-    "@types/pg": "^7.4.14",
+    "@types/compression": "^1.0.1",
+    "@types/connect-redis": "0.0.13",
+    "@types/dotenv": "^8.2.0",
+    "@types/express-rate-limit": "^3.3.3",
+    "@types/express-session": "^1.15.15",
+    "@types/helmet": "0.0.45",
+    "@types/jest": "^24.0.23",
+    "@types/jquery": "^3.3.31",
+    "@types/morgan": "^1.7.37",
+    "@types/node": "^12.12.7",
+    "@types/passport": "^1.0.1",
+    "@types/passport-google-oauth": "^1.0.41",
+    "@types/passport-jwt": "^3.0.3",
+    "@types/pg": "^7.11.2",
+    "@types/redis": "^2.8.14",
+    "@types/sqlite3": "^3.1.5",
     "@types/supertest": "^2.0.8",
-    "codelyzer": "^5.1.0",
-    "cypress": "^3.4.1",
-    "husky": "^3.0.0",
-    "jest": "^24.8.0",
+    "codelyzer": "^5.2.0",
+    "cypress": "^3.6.1",
+    "husky": "^3.0.9",
+    "jest": "^24.9.0",
     "jest-preset-angular": "^7.1.1",
-    "lint-staged": "^9.2.1",
+    "lint-staged": "^9.4.3",
     "npm-check": "^5.9.0",
-    "prettier": "^1.18.2",
-    "snyk": "^1.193.0",
+    "prettier": "^1.19.1",
+    "snyk": "^1.247.1",
     "supertest": "^4.0.2",
-    "ts-jest": "^24.0.2",
-    "ts-node": "^8.3.0",
-    "tslint": "^5.18.0",
+    "ts-jest": "^24.1.0",
+    "ts-node": "^8.5.2",
+    "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0",
+    "tslint-sonarts": "^1.9.0",
     "typescript": "^3.4.5"
   },
   "engines": {
