@@ -1,67 +1,32 @@
-import { DynamicModule, Global, Module, Provider } from '@nestjs/common';
-import { DATABASE_MODULE_OPTIONS } from './database.constants';
-import { createDatabaseProvider } from './database.provider';
+import { AsyncModuleConfig } from '@golevelup/nestjs-modules';
+import { DynamicModule, Module } from '@nestjs/common';
+import { LoggerModule } from '../logger/logger.module';
+import { DatabaseCoreModule } from './database-core.module';
+import { createDatabaseFeatureProvider } from './database.provider';
 import { DatabaseService } from './database.service';
-import {
-  DatabaseModuleAsyncOptions,
-  DatabaseModuleOptions,
-  DatabaseOptionsFactory,
-} from './interfaces/database-options.interface';
+import { DatabaseModuleOptions } from './interfaces/database-options.interface';
+import { DatabaseFeatureOptions } from './interfaces/database.interface';
 
-@Global()
 @Module({
-  providers: [DatabaseService],
-  exports: [DatabaseService],
+  imports: [LoggerModule.forFeature({ context: DatabaseService.name })],
 })
 export class DatabaseModule {
   static forRoot(options: DatabaseModuleOptions): DynamicModule {
+    return DatabaseCoreModule.forRoot(DatabaseCoreModule, options);
+  }
+
+  static forRootAsync(
+    options: AsyncModuleConfig<DatabaseModuleOptions>,
+  ): DynamicModule {
+    return DatabaseCoreModule.forRootAsync(DatabaseCoreModule, options);
+  }
+
+  static forFeature(options: DatabaseFeatureOptions): DynamicModule {
     return {
       module: DatabaseModule,
-      providers: createDatabaseProvider(options),
-    };
-  }
-
-  static forRootAsync(options: DatabaseModuleAsyncOptions): DynamicModule {
-    return {
-      module: DatabaseModule,
-      imports: options.imports || [],
-      providers: this.createAsyncProviders(options),
-    };
-  }
-
-  private static createAsyncProviders(
-    options: DatabaseModuleAsyncOptions,
-  ): Provider[] {
-    if (options.useExisting || options.useFactory) {
-      return [this.createAsyncOptionsProvider(options)];
-    }
-    if (options.useClass) {
-      return [
-        this.createAsyncOptionsProvider(options),
-        {
-          provide: options.useClass,
-          useClass: options.useClass,
-        },
-      ];
-    }
-    throw new Error('Invalid DatabaseModule configuration.');
-  }
-
-  private static createAsyncOptionsProvider(
-    options: DatabaseModuleAsyncOptions,
-  ): Provider {
-    if (options.useFactory) {
-      return {
-        provide: DATABASE_MODULE_OPTIONS,
-        useFactory: options.useFactory,
-        inject: options.inject || [],
-      };
-    }
-    return {
-      provide: DATABASE_MODULE_OPTIONS,
-      useFactory: async (optionsFactory: DatabaseOptionsFactory) =>
-        await optionsFactory.createDatabaseOptions(),
-      inject: [options.useExisting || options.useClass || ''],
+      imports: [DatabaseCoreModule.Deferred],
+      providers: [...createDatabaseFeatureProvider(options), DatabaseService],
+      exports: [DatabaseService],
     };
   }
 }
