@@ -1,14 +1,15 @@
-import { Provider } from '@nestjs/common';
-import { DATABASE_MODULE_OPTIONS } from './database.constants';
-import { DatabaseModuleOptions } from './interfaces/database-options.interface';
+import { Provider, Scope } from '@nestjs/common';
+import { DATABASE_FEATURE } from './database.constants';
+import { DatabaseFeatureOptions } from './interfaces/database.interface';
 
-export function createDatabaseProvider(
-  options: DatabaseModuleOptions,
+export function createDatabaseFeatureProvider(
+  options: DatabaseFeatureOptions,
 ): Provider[] {
   return [
     {
-      provide: DATABASE_MODULE_OPTIONS,
-      useValue: options || {},
+      provide: DATABASE_FEATURE,
+      useValue: options,
+      scope: Scope.TRANSIENT,
     },
   ];
 }
