@@ -1,18 +1,9 @@
 import { ExecutionContext, Injectable } from '@nestjs/common';
 import { GqlExecutionContext } from '@nestjs/graphql';
-import { AuthGuard } from '@nestjs/passport';
+import { GoogleGuard } from './google.guard';
 
 @Injectable()
-export class GqlAuthGuard extends AuthGuard('jwt') {
-  constructor() {
-    super();
-  }
-
-  canActivate(context: ExecutionContext) {
-    // added to be able to extend the authentication logic if need be
-    return super.canActivate(context);
-  }
-
+export class GqlAuthGuard extends GoogleGuard {
   getRequest(context: ExecutionContext) {
     const ctx = GqlExecutionContext.create(context);
     return ctx.getContext().req;
