@@ -1,18 +1,39 @@
 import { INestApplication, ValidationPipe } from '@nestjs/common';
 import * as compression from 'compression';
+import * as store from 'connect-redis';
 import * as rateLimiter from 'express-rate-limit';
+import * as session from 'express-session';
 import * as helmet from 'helmet';
 import * as morgan from 'morgan';
+import * as passport from 'passport';
+import * as redis from 'redis';
 import { ConfigService } from './app/config/config.service';
-import { MyLogger } from './app/logger/logger.service';
+import { LoggerService } from './app/logger/logger.service';
 
-export function configure(app: INestApplication, config: ConfigService): void {
-  const morganFormat = config.isProd() ? 'combined' : 'dev';
+const RedisStore = store(session);
+
+export function configure(
+  app: INestApplication,
+  config: ConfigService,
+  logger: LoggerService,
+): void {
   app.use(
-    morgan(morganFormat, {
-      skip: (req: any, res: any) => config.isProd() && req.statusCode < 400,
+    session({
+      store: new RedisStore({
+        client: redis.createClient({
+          url: config.getRedisUrl(),
+        }),
+      }),
+      secret: config.getSessionSecret(),
+      resave: false,
+      saveUninitialized: false,
+    }),
+    morgan(config.getMorganString(), {
+      skip: (req: any, res: any) =>
+        (config.isProd() && req.statusCode < 400) ||
+        req.url.includes('callback'),
       stream: {
-        write: (value: string) => MyLogger.log(value.trim(), 'Morgan'),
+        write: (value: string) => logger.log(value.trim(), 'Morgan'),
       },
     }),
     helmet(),
@@ -21,8 +42,10 @@ export function configure(app: INestApplication, config: ConfigService): void {
       windowMs: 10 * 60 * 1000,
       max: config.getRateLimit(),
     }),
+    passport.initialize(),
+    passport.session(),
   );
-  app.setGlobalPrefix(config.get('GLOBAL_PREFIX'));
+  app.setGlobalPrefix(config.getGlobalPrefix());
   app.useGlobalPipes(new ValidationPipe());
-  MyLogger.log('Application Configuration complete', 'ApplicationConfig');
+  logger.log('Application Configuration complete', 'ApplicationConfig');
 }
